Extract auth guard
