+++ mikado-app/src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import * as React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import Plaza from './pages/Graph/Plaza';
+import GraphPage from "./pages/Graph/GraphPage";
 import SignIn from './pages/SignIn/SignIn';
 import { FirebaseContextProvider } from './context/FirebaseContext';
 
@@ -11,7 +11,7 @@ function App() {
       <div>
         <Routes>
           <Route path='/' element={<SignIn />} />
-          <Route path='/graph' element={<Plaza />} />
+          <Route path='/graph' element={<GraphPage />} />
         </Routes>
       </div>
     </FirebaseContextProvider>
