+++ mikado-app/src/pages/Graph/GraphPage.js
@@ -1,0 +1,8 @@
+import AuthenticatedUidWrapper from "../../context/AuthenticatedUidWrapper";
+import Plaza from "./Plaza";
+
+function GraphPage() {
+  return <AuthenticatedUidWrapper wrapped={Plaza} />
+}
+
+export default GraphPage;
